requirements.txt
Flask-SocketIO
python-engineio
python-socketio
opencv-python
pyserial
flask
websockets
aiohttp